+++ KroesSupermarketMod/ConfigManager.cs
@@ -11,12 +11,14 @@ namespace KroesSupermarketMod
         private ConfigEntry<bool> chatCommandsCfg;
         private ConfigEntry<bool> euroSymbolCfg;
         private ConfigEntry<bool> throwBoxesCfg;
+        private ConfigEntry<bool> thirdPersonCameraCfg;
         internal bool customNpcHitNotifications => customNpcHitNotificationCfg.Value;
         internal bool doublePriceGun => doublePriceGunCfg.Value;
         internal bool franchiseProgressBar => franchiseProgressBarCfg.Value;
         internal bool chatCommands => chatCommandsCfg.Value;
         internal bool euroSymbol => euroSymbolCfg.Value;
         internal bool throwBoxes => throwBoxesCfg.Value;
+        internal bool thirdPersonCamera => thirdPersonCameraCfg.Value;
         public ConfigManager(ConfigFile config)
         {
             customNpcHitNotificationCfg = config.Bind("General", "CustomNpcHitNotifications", false, "Enable custom voice lines (inside jokes).");
@@ -25,6 +27,7 @@ namespace KroesSupermarketMod
             chatCommandsCfg = config.Bind("General", "ChatCommands", true, "Enable chat commands feature.");
             euroSymbolCfg = config.Bind("General", "EuroSymbol", false, "Replace dollar sign with euro.");
             throwBoxesCfg = config.Bind("General", "ThrowBoxes", true, "Throw boxes instead of dropping.");
+            thirdPersonCameraCfg = config.Bind("General", "ThirdPersonCamera", true, "Allow camera switching with mouse scroll wheel.");
         }
 
     }